Move all exceptions into exceptions.py
