+++ openff/toolkit/utils/exceptions.py
@@ -24,22 +24,29 @@ __all__ = (
 
 
 class MessageException(Exception):
-    """A base class for exceptions that print out a string given in their constructor"""
+    """DEPRECATED: A base class for exceptions that print out a string given in their constructor"""
 
-    def __init__(self, msg):
-        super().__init__(self, msg)
-        self.msg = msg
+    import warnings
 
-    def __str__(self):
-        return self.msg
+    # Must "turn on" DeprecationWawrning, see https://stackoverflow.com/a/60486394/4248961
+    warnings.filterwarnings("default", category=DeprecationWarning, module=__name__)
+
+    warnings.warn(
+        "MessageException is deprecated and will be removed in a future release of the OpenFF Toolkit."
+        "All custom exceptions now inherit from OpenFFToolkitException, which should be used as a "
+        "replacement for MessageException. (`from openff.toolkit.utils.exceptions import OpenFFToolkitException`",
+        DeprecationWarning,
+    )
+
+
+class OpenFFToolkitException(BaseException):
+    """Base exception for custom exceptions raised by the OpenFF Toolkit"""
 
 
 class IncompatibleUnitError(MessageException):
     """
     Exception for when a parameter is in the wrong units for a ParameterHandler's unit system
     """
-
-    pass
 
 
 class MissingDependencyError(MessageException):
@@ -111,5 +118,165 @@ class AntechamberNotFoundError(MessageException):
     """The antechamber executable was not found"""
 
 
-class ParseError(MessageException, ValueError):
+class ParseError(ValueError):
     """The record couple not be parsed into the given format"""
+
+
+class InvalidConformerError(Exception):
+    """
+    This error is raised when the conformer added to the molecule
+    has a different connectivity to that already defined.
+    or anyother conformer related issues.
+    """
+
+
+class SmilesParsingError(Exception):
+    """
+    This error is rasied when parsing a smiles string results in an error.
+    """
+
+
+class NotAttachedToMoleculeError(MessageException):
+    """Exception for when a component does not belong to a Molecule object, but is queried"""
+
+
+class DuplicateUniqueMoleculeError(MessageException):
+    """
+    Exception for when the user provides indistinguishable unique molecules when trying to identify atoms from a PDB
+    """
+
+
+class NotBondedError(MessageException):
+    """
+    Exception for when a function requires a bond between two atoms, but none is present
+    """
+
+
+class InvalidBoxVectorsError(MessageException):
+    """
+    Exception for setting invalid box vectors
+    """
+
+
+class InvalidPeriodicityError(MessageException):
+    """
+    Exception for setting invalid periodicity
+    """
+
+
+class MissingUniqueMoleculesError(MessageException):
+    """
+    Exception for a when unique_molecules is required but not found
+    """
+
+
+class SMIRKSMismatchError(MessageException):
+    """
+    Exception for cases where smirks are inappropriate
+    for the environment type they are being parsed into
+    """
+
+
+class SMIRKSParsingError(MessageException):
+    """
+    Exception for when SMIRKS are not parseable for any environment
+    """
+
+
+class ParameterHandlerRegistrationError(MessageException):
+    """
+    Exception for errors in ParameterHandler registration
+    """
+
+
+class SMIRNOFFVersionError(MessageException):
+    """
+    Exception thrown when an incompatible SMIRNOFF version data structure in attempted to be read.
+    """
+
+
+class SMIRNOFFAromaticityError(MessageException):
+    """
+    Exception thrown when an incompatible SMIRNOFF aromaticity model is checked for compatibility.
+    """
+
+
+class ParseError(MessageException):
+    """
+    Error for when a SMIRNOFF data structure is not parseable by a ForceField
+    """
+
+
+class PartialChargeVirtualSitesError(MessageException):
+    """
+    Exception thrown when partial charges cannot be computed for a Molecule because the ForceField applies virtual sites.
+    """
+
+
+class SMIRNOFFSpecError(MessageException):
+    """
+    Exception for when data is noncompliant with the SMIRNOFF data specification.
+    """
+
+
+class SMIRNOFFSpecUnimplementedError(MessageException):
+    """
+    Exception for when a portion of the SMIRNOFF specification is not yet implemented.
+    """
+
+
+class FractionalBondOrderInterpolationMethodUnsupportedError(MessageException):
+    """
+    Exception for when an unsupported fractional bond order interpolation assignment method is called.
+    """
+
+
+class NotEnoughPointsForInterpolationError(MessageException):
+    """Exception for when less than two points are provided for interpolation"""
+
+
+class IncompatibleParameterError(MessageException):
+    """
+    Exception for when a set of parameters is scientifically/technically incompatible with another
+    """
+
+
+class UnassignedValenceParameterException(Exception):
+    """Exception raised when there are valence terms for which a ParameterHandler can't find parameters."""
+
+
+class UnassignedBondParameterException(UnassignedValenceParameterException):
+    """Exception raised when there are bond terms for which a ParameterHandler can't find parameters."""
+
+
+class UnassignedAngleParameterException(UnassignedValenceParameterException):
+    """Exception raised when there are angle terms for which a ParameterHandler can't find parameters."""
+
+
+class UnassignedProperTorsionParameterException(UnassignedValenceParameterException):
+    """Exception raised when there are proper torsion terms for which a ParameterHandler can't find parameters."""
+
+
+class UnassignedMoleculeChargeException(Exception):
+    """Exception raised when no charge method is able to assign charges to a molecule."""
+
+
+class NonintegralMoleculeChargeException(Exception):
+    """Exception raised when the partial charges on a molecule do not sum up to its formal charge."""
+
+
+class DuplicateParameterError(MessageException):
+    """Exception raised when trying to add a ParameterType that already exists"""
+
+
+class ParameterLookupError(MessageException):
+    """Exception raised when something goes wrong in a parameter lookup in
+    ParameterHandler.__getitem__"""
+
+
+class DuplicateVirtualSiteTypeException(Exception):
+    """Exception raised when trying to register two different virtual site classes with the same 'type'"""
+
+
+class CallbackRegistrationError(TypeError):
+    """Error raised when callback registration fails."""
