+++ openff/toolkit/utils/toolkits.py
@@ -20,51 +20,25 @@ Currently supported toolkits:
    * Change global variables from _INSTALLED to _AVAILABLE
 
 """
-__all__ = (
-    # constants
+__all__ = [
     "DEFAULT_AROMATICITY_MODEL",
     "ALLOWED_AROMATICITY_MODELS",
     "DEFAULT_FRACTIONAL_BOND_ORDER_MODEL",
     "ALLOWED_FRACTIONAL_BOND_ORDER_MODELS",
     "DEFAULT_CHARGE_MODEL",
     "ALLOWED_CHARGE_MODELS",
-    # exceptions and warnings
-    "MessageException",
-    "IncompatibleUnitError",
-    "MissingDependencyError",
-    "MissingPackageError",
-    "ToolkitUnavailableException",
-    "LicenseError",
-    "InvalidToolkitError",
-    "InvalidToolkitRegistryError",
-    "UndefinedStereochemistryError",
-    "GAFFAtomTypeWarning",
-    "ChargeMethodUnavailableError",
-    "IncorrectNumConformersError",
-    "IncorrectNumConformersWarning",
-    "ChargeCalculationError",
-    "InvalidIUPACNameError",
-    "AntechamberNotFoundError",
-    "ParseError",
-    # base_wrapper
     "ToolkitWrapper",
-    # builtin_wrapper
     "BuiltInToolkitWrapper",
-    # openeye_wrapper
     "OpenEyeToolkitWrapper",
-    # rdkit_wrapper
     "RDKitToolkitWrapper",
-    # ambertools_wrapper
     "AmberToolsToolkitWrapper",
-    # toolkit_registry
     "ToolkitRegistry",
-    # in this module
     "GLOBAL_TOOLKIT_REGISTRY",
     "OPENEYE_AVAILABLE",
     "RDKIT_AVAILABLE",
     "AMBERTOOLS_AVAILABLE",
     "BASIC_CHEMINFORMATICS_TOOLKITS",
-)
+]
 
 
 # =============================================================================================
@@ -72,6 +46,31 @@ __all__ = (
 # =============================================================================================
 
 import logging
+
+from openff.toolkit.utils.exceptions import (
+    AntechamberNotFoundError,
+    ChargeCalculationError,
+    ChargeMethodUnavailableError,
+    GAFFAtomTypeWarning,
+    IncompatibleUnitError,
+    IncorrectNumConformersError,
+    IncorrectNumConformersWarning,
+    InvalidIUPACNameError,
+    InvalidToolkitError,
+    InvalidToolkitRegistryError,
+    LicenseError,
+    MissingDependencyError,
+    MissingPackageError,
+    ParseError,
+    ToolkitUnavailableException,
+    UndefinedStereochemistryError,
+)
+from openff.toolkit.utils.openeye_wrapper import (
+    OpenEyeToolkitWrapper,
+    requires_openeye_module,
+)
+from openff.toolkit.utils.rdkit_wrapper import RDKitToolkitWrapper
+from openff.toolkit.utils.toolkit_registry import ToolkitRegistry
 
 from .ambertools_wrapper import AmberToolsToolkitWrapper
 from .base_wrapper import ToolkitWrapper
@@ -84,28 +83,6 @@ from .constants import (
     DEFAULT_CHARGE_MODEL,
     DEFAULT_FRACTIONAL_BOND_ORDER_MODEL,
 )
-from .exceptions import (
-    AntechamberNotFoundError,
-    ChargeCalculationError,
-    ChargeMethodUnavailableError,
-    GAFFAtomTypeWarning,
-    IncompatibleUnitError,
-    IncorrectNumConformersError,
-    IncorrectNumConformersWarning,
-    InvalidIUPACNameError,
-    InvalidToolkitError,
-    InvalidToolkitRegistryError,
-    LicenseError,
-    MessageException,
-    MissingDependencyError,
-    MissingPackageError,
-    ParseError,
-    ToolkitUnavailableException,
-    UndefinedStereochemistryError,
-)
-from .openeye_wrapper import OpenEyeToolkitWrapper, requires_openeye_module
-from .rdkit_wrapper import RDKitToolkitWrapper
-from .toolkit_registry import ToolkitRegistry
 
 # =============================================================================================
 # CONFIGURE LOGGER
@@ -165,7 +142,7 @@ if (
     )
     == 0
 ):
-    from .utils import all_subclasses
+    from openff.toolkit.utils import all_subclasses
 
     msg = "WARNING: No basic cheminformatics toolkits are available.\n"
     msg += "At least one basic toolkit is required to handle SMARTS matching and file I/O. \n"
