+++ openff/toolkit/utils/utils.py
@@ -38,8 +38,10 @@ import logging
 
 from simtk import unit
 
-from .exceptions import MessageException  # Needed for backward compatibility
-from .exceptions import IncompatibleUnitError, MissingDependencyError
+from openff.toolkit.utils.exceptions import (
+    IncompatibleUnitError,
+    MissingDependencyError,
+)
 
 # =============================================================================================
 # CONFIGURE LOGGER
